download fim from tools/fim_manager.py
